Api_Reconocimiento_Texto/requirements.txt
Flask==2.0.3
pytesseract==0.3.9
Pillow==9.2.0
